Refactor e2e test
