+++ deploy/olm-catalog/common-service-operator/0.0.3/common-service-operator.v0.0.3.clusterserviceversion.yaml
@@ -6,6 +6,35 @@ metadata:
       [
         {
           "apiVersion": "operator.ibm.com/v1alpha1",
+          "kind": "CommonServiceConfig",
+          "metadata": {
+            "name": "common-service"
+          },
+          "spec": {
+            "services": [
+              {
+                "name": "jenkins",
+                "spec": {
+                  "jenkins": {
+                    "service": {
+                      "port": 8081
+                    }
+                  }
+                }
+              },
+              {
+                "name": "etcd",
+                "spec": {
+                  "etcdCluster": {
+                    "size": 1
+                  }
+                }
+              }
+            ]
+          }
+        },
+        {
+          "apiVersion": "operator.ibm.com/v1alpha1",
           "kind": "CommonServiceSet",
           "metadata": {
             "name": "common-service"
@@ -52,35 +81,6 @@ metadata:
                 "targetNamespaces": [
                   "etcd-operator"
                 ]
-              }
-            ]
-          }
-        },
-        {
-          "apiVersion": "operator.ibm.com/v1alpha1",
-          "kind": "CommonServiceConfig",
-          "metadata": {
-            "name": "common-service"
-          },
-          "spec": {
-            "services": [
-              {
-                "name": "jenkins",
-                "spec": {
-                  "jenkins": {
-                    "service": {
-                      "port": 8081
-                    }
-                  }
-                }
-              },
-              {
-                "name": "etcd",
-                "spec": {
-                  "etcdCluster": {
-                    "size": 1
-                  }
-                }
               }
             ]
           }
